+++ pkg/providers/gtm/testdata/domain_with_sign_and_serve/import.sh
@@ -1,0 +1,2 @@
+terraform init
+terraform import akamai_gtm_domain.test_name "test.name.akadns.net"