+++ pkg/providers/gtm/testdata/with_multiline/domain.tf
@@ -32,4 +32,5 @@ EOT
   cname_coalescing_enabled  = true
   load_feedback             = true
   end_user_mapping_enabled  = false
+  sign_and_serve            = false
 }