+++ pkg/providers/gtm/testdata/with_multiline/properties.tf
@@ -29,6 +29,7 @@ EOT
     http_error3xx                    = true
     http_error4xx                    = true
     http_error5xx                    = true
+    pre_2023_security_posture        = false
     disabled                         = false
     test_object_protocol             = "HTTP"
     test_object_port                 = 80